+++ source/ftrack_connect_pipeline/__init__.py
@@ -1,5 +1,5 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2016 ftrack
 
-from ._version import __version__
 
+from ._version import __version__