+++ test/interactive/resource/assets/camera/camera_asset.py
@@ -8,10 +8,10 @@ IDENTIFIER = 'camera'
 class PublishCamera(ftrack_connect_pipeline.asset.PyblishAsset):
     '''Handle publish of maya camera.'''
 
-    def get_publish_items(self, publish_data):
+    def get_publish_items(self):
         '''Return list of items that can be published.'''
         options = []
-        for instance in publish_data:
+        for instance in self.pyblish_context:
             if instance.data['family'] in ('camera', ):
                 options.append({
                     'label': instance.name,
@@ -21,10 +21,10 @@ class PublishCamera(ftrack_connect_pipeline.asset.PyblishAsset):
 
         return options
 
-    def get_item_options(self, publish_data, name):
+    def get_item_options(self, name):
         '''Return options for publishable item with *name*.'''
         options = []
-        for instance in publish_data:
+        for instance in self.pyblish_context:
             if instance.data['family'] in ('camera', ):
                 options = [{
                     'type': 'boolean',
@@ -38,9 +38,9 @@ class PublishCamera(ftrack_connect_pipeline.asset.PyblishAsset):
 
         return options
 
-    def get_options(self, publish_data):
+    def get_options(self):
         '''Return global options for publishing.'''
-        options = super(PublishCamera, self).get_options(publish_data)
+        options = super(PublishCamera, self).get_options()
         options.append({
             'type': 'group',
             'name': 'camera_options',
@@ -58,6 +58,14 @@ class PublishCamera(ftrack_connect_pipeline.asset.PyblishAsset):
         return options
 
 
+def create_asset_publish():
+    '''Return publish camera.'''
+    return PublishCamera(
+        description='publish maya cameras to ftrack.',
+        asset_type_short='camera'
+    )
+
+
 def register(session):
     '''Subscribe to *session*.'''
     if not isinstance(session, ftrack_api.Session):
@@ -65,11 +73,9 @@ def register(session):
 
     camera_asset = ftrack_connect_pipeline.asset.Asset(
         identifier=IDENTIFIER,
-        publish_asset=PublishCamera(
-            label='Camera',
-            description='publish maya cameras to ftrack.',
-            icon='http://www.clipartbest.com/cliparts/9Tp/erx/9Tperxqrc.png'
-        )
+        label='Camera',
+        icon='http://www.clipartbest.com/cliparts/9Tp/erx/9Tperxqrc.png',
+        create_asset_publish=create_asset_publish
     )
     # Register camera asset on session. This makes sure that discover is called
     # for import and publish.