+++ doc/release/release_notes.rst
@@ -13,7 +13,17 @@ Release Notes
 
         Added thumbnail selector.
 
+.. release:: 0.1.3
+    :date: 2016-11-22
+
+    .. change:: changed
+        :tags: Publish, Pyblish
+
+        Centralised shared pyblish plugins.
+
 .. release:: 0.1.2
+    :date: 2016-11-18
+
     .. change:: new
         :tags: Widget